pyproject.toml
[project]
name = "balt-dom-bot"
version = "0.1.0"
description = "AI-бот для чатов ЖК «Балтийский Дом» в мессенджере Max"
requires-python = ">=3.11"
authors = [{ name = "Балтийский Дом" }]
dependencies = [
    "maxapi==1.0.0",
    "pydantic>=2.7,<3",
    "pydantic-settings>=2.4,<3",
    "PyYAML>=6.0,<7",
    "structlog>=24.1,<26",
    "httpx>=0.27,<1",
    "aiosqlite>=0.20,<1",
    "fastapi>=0.110,<1",
    "uvicorn[standard]>=0.30,<1",
    "Jinja2>=3.1,<4",
    "python-multipart>=0.0.9,<1",
    "PyJWT>=2.8,<3",
    "bcrypt>=4.0,<5",
]

[project.optional-dependencies]
webhook = ["fastapi>=0.110,<1", "uvicorn[standard]>=0.30,<1"]
dev = ["pytest>=8", "pytest-asyncio>=0.23", "ruff>=0.6", "mypy>=1.10"]

[project.scripts]
balt-dom-bot = "balt_dom_bot.__main__:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/balt_dom_bot"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "RUF", "SIM"]
ignore = ["E501"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
